requirements.txt
python-dotenv>=1.0,<2.0
Flask>=3.0,<4.0
Werkzeug>=3.0,<4.0
requests>=2.31,<3.0
Flask-Limiter>=3.8,<4.0
pypdf
PyMuPDF
gunicorn==25.1.0
tzdata>=2024.1
openpyxl>=3.1,<4.0
python-docx>=1.1,<2.0
